refactor/#61: AuthenticationCodeRepository 추상화
