+++ src/main/java/com/tookscan/tookscan/security/application/service/ReadSerialIdAndProviderService.java
@@ -1,15 +1,13 @@
 package com.tookscan.tookscan.security.application.service;
 
 import com.tookscan.tookscan.account.repository.UserRepository;
-import com.tookscan.tookscan.core.exception.error.ErrorCode;
-import com.tookscan.tookscan.core.exception.type.CommonException;
 import com.tookscan.tookscan.security.application.dto.request.ReadSerialIdAndProviderRequestDto;
 import com.tookscan.tookscan.security.application.dto.response.ReadSerialIdAndProviderResponseDto;
 import com.tookscan.tookscan.security.application.usecase.ReadSerialIdAndProviderUseCase;
 import com.tookscan.tookscan.security.domain.mysql.Account;
 import com.tookscan.tookscan.security.domain.redis.AuthenticationCode;
 import com.tookscan.tookscan.security.domain.service.AuthenticationCodeService;
-import com.tookscan.tookscan.security.repository.redis.AuthenticationCodeRepository;
+import com.tookscan.tookscan.security.repository.AuthenticationCodeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,8 +26,7 @@ public class ReadSerialIdAndProviderService implements ReadSerialIdAndProviderUs
     public ReadSerialIdAndProviderResponseDto execute(ReadSerialIdAndProviderRequestDto requestDto) {
 
         // 인증 코드 조회
-        AuthenticationCode authenticationCode = authenticationCodeRepository.findById(requestDto.phoneNumber())
-                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_RESOURCE));
+        AuthenticationCode authenticationCode = authenticationCodeRepository.findByIdOrElseThrow(requestDto.phoneNumber());
 
         // 인증 코드 검증 확인
         authenticationCodeService.validateAuthenticationCode(authenticationCode);
