+++ src/main/java/com/tookscan/tookscan/security/application/service/ValidateAuthenticationCodeService.java
@@ -5,7 +5,7 @@ import com.tookscan.tookscan.core.exception.type.CommonException;
 import com.tookscan.tookscan.security.application.dto.request.ValidateAuthenticationCodeRequestDto;
 import com.tookscan.tookscan.security.application.usecase.ValidateAuthenticationCodeUseCase;
 import com.tookscan.tookscan.security.domain.redis.AuthenticationCode;
-import com.tookscan.tookscan.security.repository.redis.AuthenticationCodeRepository;
+import com.tookscan.tookscan.security.repository.AuthenticationCodeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -22,8 +22,7 @@ public class ValidateAuthenticationCodeService implements ValidateAuthentication
     public void execute(ValidateAuthenticationCodeRequestDto requestDto) {
 
         // 해당 전화번호로 발급된 인증코드 조회
-        AuthenticationCode authenticationCode = authenticationCodeRepository.findById(requestDto.phoneNumber())
-                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_RESOURCE));
+        AuthenticationCode authenticationCode = authenticationCodeRepository.findByIdOrElseThrow(requestDto.phoneNumber());
 
         // 인증코드 일치 여부 확인
         if (!bCryptPasswordEncoder.matches(requestDto.authenticationCode(), authenticationCode.getValue())) {
