+++ src/main/java/com/tookscan/tookscan/security/repository/redis/AuthenticationCodeRedisRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthenticationCodeRepository extends CrudRepository<AuthenticationCode, String>{
+public interface AuthenticationCodeRedisRepository extends CrudRepository<AuthenticationCode, String>{
 }
